feat: Add device info characteristic and UI modal for device status (#60)
* feat: Add device info characteristic and UI modal for displaying device status

- Implemented a new Device Info characteristic in BLEManager for providing battery level, RSSI, SNR, and LoRa configuration.
- Updated BLEManager to handle device info requests and serialize data for transmission.
- Created a new DeviceInfoModal component to display device information in the PWA.
- Modified ConnectionStatus component to include an info button for requesting device info.
- Removed battery level monitoring from the PWA, as it is now part of the device info characteristic.
- Updated BleService to read device info from the new characteristic and handle requests from the UI.

* fix(ChatViewModelConnectionTest): update connection status text to remove error icon

* feat(BLEManager): add updateDeviceInfo method to refresh device info characteristic

* feat(BLEManager): implement updateDeviceInfo method and refactor onRead callback

* feat(ConnectionStatus): enhance layout for small screens and add device info button

* feat(BLEManager): update onRead method to include characteristic and connection info parameters
feat(LEDManager): change heartbeat and blink duration parameters to unsigned long type
fix(vscode): correct formatting in extensions.json recommendations

+++ firmware/include/common/LEDManager.h
@@ -21,7 +21,7 @@ public:
      * @param hbIntervalMs Heartbeat interval in milliseconds (default: 1000)
      * @param hbDurationMs Heartbeat blink duration in milliseconds (default: 50)
      */
-    explicit LEDManager(int pin, unsigned long hbIntervalMs = 1000, int hbDurationMs = 50)
+    explicit LEDManager(int pin, unsigned long hbIntervalMs = 1000, unsigned long hbDurationMs = 50)
         : ledPin(pin), blinkActive(false), blinkCount(0),
           blinkTarget(0), blinkDuration(50), blinkDelay(200),
           lastStateChange(0), ledState(false),
@@ -43,7 +43,7 @@ public:
      * @param duration Time LED stays on in each blink in milliseconds (default: 50)
      * @param delayBetween Delay between blinks in milliseconds (default: 200)
      */
-    void blink(int times = 1, int duration = 50, int delayBetween = 200)
+    void blink(int times = 1, unsigned long duration = 50, unsigned long delayBetween = 200)
     {
         blinkActive = true;
         blinkTarget = times;
@@ -128,13 +128,13 @@ private:
     bool blinkActive;
     int blinkCount;
     int blinkTarget;
-    int blinkDuration;
-    int blinkDelay;
+    unsigned long blinkDuration;
+    unsigned long blinkDelay;
     unsigned long lastStateChange;
     bool ledState;
     unsigned long lastHeartbeat;
     unsigned long heartbeatIntervalMs;
-    int heartbeatDurationMs;
+    unsigned long heartbeatDurationMs;
 };
 
 #endif // LED_MANAGER_H
